app/services/image_service.py
"""
Image Processing Service - Architecture Ready
This service is designed to handle image processing for the Vexere AI system.
Currently contains the architecture and interfaces, ready for implementation.
"""

from typing import Dict, Any, Optional
import base64
from PIL import Image
import io

class ImageProcessingService:
    """
    Service for processing images in customer service context
    Architecture ready for OCR, document analysis, and RAG integration
    """
    
    def __init__(self):
        self.initialized = False
        self.ocr_engine = None  # Ready for Tesseract, EasyOCR, or cloud OCR
        self.image_analysis_model = None  # Ready for vision models
    
    def initialize(self):
        """
        Initialize image processing components
        """
        # TODO: Initialize OCR engine (Tesseract, EasyOCR, etc.)
        # TODO: Initialize image analysis model (CLIP, BLIP, etc.)
        # TODO: Setup image preprocessing pipeline
        self.initialized = True
        return True
    
    def preprocess_image(self, image_data: bytes) -> Dict[str, Any]:
        """
        Preprocess uploaded image
        """
        try:
            # Convert bytes to PIL Image
            image = Image.open(io.BytesIO(image_data))
            
            # Basic preprocessing
            processed_info = {
                "format": image.format,
                "size": image.size,
                "mode": image.mode,
                "processed": True
            }
            
            # TODO: Add image enhancement, noise reduction, etc.
            
            return processed_info
        except Exception as e:
            return {"error": f"Image preprocessing failed: {str(e)}"}
    
    def extract_text_from_image(self, image_data: bytes) -> Dict[str, Any]:
        """
        Extract text from image using OCR
        """
        if not self.initialized:
            return {"error": "Image service not initialized"}
        
        try:
            # Preprocess image
            preprocess_result = self.preprocess_image(image_data)
            if "error" in preprocess_result:
                return preprocess_result
            
            # TODO: Implement OCR processing
            # Example structure:
            # ocr_result = self.ocr_engine.readtext(image_data)
            # extracted_text = " ".join([item[1] for item in ocr_result])
            
            # Mock result for architecture demonstration
            extracted_text = "Mã đặt chỗ: VX001234\nHành khách: Nguyễn Văn A\nChuyến bay: VJ123"
            confidence = 0.95
            
            return {
                "extracted_text": extracted_text,
                "confidence": confidence,
                "processing_time": 0.5,
                "status": "success"
            }
        except Exception as e:
            return {"error": f"OCR processing failed: {str(e)}"}
    
    def analyze_image_content(self, image_data: bytes) -> Dict[str, Any]:
        """
        Analyze image content for context understanding
        """
        if not self.initialized:
            return {"error": "Image service not initialized"}
        
        try:
            # TODO: Implement image content analysis
            # - Document type detection (boarding pass, ticket, etc.)
            # - Object detection (airplane, airport, etc.)
            # - Scene understanding
            
            # Mock result for architecture demonstration
            analysis_result = {
                "document_type": "boarding_pass",
                "detected_objects": ["text", "barcode", "logo"],
                "confidence": 0.88,
                "suggested_actions": ["extract_booking_info", "verify_document"]
            }
            
            return analysis_result
        except Exception as e:
            return {"error": f"Image analysis failed: {str(e)}"}
    
    def process_image_for_rag(self, image_data: bytes) -> Dict[str, Any]:
        """
        Process image and prepare for RAG integration
        """
        try:
            # Extract text
            ocr_result = self.extract_text_from_image(image_data)
            if "error" in ocr_result:
                return ocr_result
            
            # Analyze content
            analysis_result = self.analyze_image_content(image_data)
            if "error" in analysis_result:
                return analysis_result
            
            # Combine results for RAG processing
            rag_input = {
                "extracted_text": ocr_result.get("extracted_text", ""),
                "document_type": analysis_result.get("document_type", "unknown"),
                "context": analysis_result.get("detected_objects", []),
                "confidence": min(ocr_result.get("confidence", 0), analysis_result.get("confidence", 0))
            }
            
            return {
                "rag_input": rag_input,
                "status": "ready_for_rag",
                "processing_time": ocr_result.get("processing_time", 0) + 0.3
            }
        except Exception as e:
            return {"error": f"Image processing for RAG failed: {str(e)}"}
    
    def validate_image_format(self, image_data: bytes) -> Dict[str, Any]:
        """
        Validate image format and size
        """
        try:
            image = Image.open(io.BytesIO(image_data))
            
            # Check format
            allowed_formats = ['JPEG', 'PNG', 'WEBP']
            if image.format not in allowed_formats:
                return {"error": f"Unsupported format: {image.format}"}
            
            # Check size
            max_size = (4096, 4096)  # 4K max
            if image.size[0] > max_size[0] or image.size[1] > max_size[1]:
                return {"error": "Image too large"}
            
            # Check file size (10MB max)
            if len(image_data) > 10 * 1024 * 1024:
                return {"error": "File too large"}
            
            return {
                "valid": True,
                "format": image.format,
                "size": image.size,
                "file_size": len(image_data)
            }
        except Exception as e:
            return {"error": f"Image validation failed: {str(e)}"}

# Future integration points:
# - Integration with cloud OCR services (Google Vision, AWS Textract)
# - Advanced image analysis with vision-language models
# - Document classification and routing
# - Multi-language OCR support
# - Image quality enhancement
