requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
sentence-transformers==2.2.2
huggingface-hub>=0.19.0,<0.20.0
faiss-cpu==1.7.4
openai==1.3.5
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2